+++ src/simple_taboo_search/clique_count.h
@@ -1,9 +1,9 @@
 #ifndef CLIQUE_COUNT_H
 #define CLIQUE_COUNT_H
 
-unsigned long int CliqueCount(int *g, int gsize);
-unsigned long int CliqueCountEdge(int *g, int gsize, int i, int j);
-unsigned long int CliqueCountAll(int *g, int gsize, unsigned long int *ecounts);
+int CliqueCount(int *g, int gsize);
+int CliqueCountEdge(int *g, int gsize, int i, int j);
+int CliqueCountAll(int *g, int gsize, int *ecounts);
 
 #endif
